Reduced branching, wasteful math operations, and unused code
Searches are about 5-10% faster. LoadSpan256 and StoreSpan256 used Unsafe.AsPointer to get a pointer to the data, but this only needs to be done a single time for each call to the various Propagate methods, rather than the many millions of times it was being called each time that a vector was loaded or stored.